ext2: Change inode from an int to long. Add support for creating an ext2 file from an inode but without an Ext2Entry.

+++ fs/src/fs/org/jnode/fs/ext2/Ext2Directory.java
@@ -214,7 +214,7 @@ public class Ext2Directory extends AbstractFSDirectory implements FSDirectoryId 
         synchronized (((Ext2FileSystem) getFileSystem()).getInodeCache()) {
             //reread the inode before synchronizing to it to make sure
             //all threads use the same instance
-            int iNodeNr = iNode.getINodeNr();
+            long iNodeNr = iNode.getINodeNr();
             iNode = ((Ext2FileSystem) getFileSystem()).getINode(iNodeNr);
 
             //lock the inode into the cache so it is not flushed before synchronizing to it
@@ -333,7 +333,7 @@ public class Ext2Directory extends AbstractFSDirectory implements FSDirectoryId 
 
     @Override
     public String getDirectoryId() {
-        return Integer.toString(iNode.getINodeNr());
+        return Long.toString(iNode.getINodeNr());
     }
 
     @Override
